Namespace function seeded_rand()
