+++ lib/puppet/functions/seeded_rand.rb
@@ -1,22 +1,14 @@
 # frozen_string_literal: true
 
-# @summary
-#   Generates a random whole number greater than or equal to 0 and less than max, using the value of seed for repeatable randomness.
+# THIS FILE WAS GENERATED BY `rake regenerate_unamespaced_shims`
+
+# @summary DEPRECATED.  Use the namespaced function [`stdlib::seeded_rand`](#stdlibseeded_rand) instead.
 Puppet::Functions.create_function(:seeded_rand) do
-  # @param max The maximum value.
-  # @param seed The seed used for repeatable randomness.
-  #
-  # @return [Integer]
-  #   A random number greater than or equal to 0 and less than max
-  dispatch :seeded_rand do
-    param 'Integer[1]', :max
-    param 'String', :seed
+  dispatch :deprecation_gen do
+    repeated_param 'Any', :args
   end
-
-  def seeded_rand(max, seed)
-    require 'digest/md5'
-
-    seed = Digest::MD5.hexdigest(seed).hex
-    Puppet::Util.deterministic_rand_int(seed, max)
+  def deprecation_gen(*args)
+    call_function('deprecation', 'seeded_rand', 'This function is deprecated, please use stdlib::seeded_rand instead.')
+    call_function('stdlib::seeded_rand', *args)
   end
 end
